rozbudowazapytanie-service/requirements.txt
# Główna biblioteka Agent Development Kit
google-adk

# Biblioteka do konfiguracji Vertex AI i komunikacji z API
google-generativeai

# Serwer webowy i klient HTTP
fastapi
uvicorn
requests
